app/requirements.txt
django-bootstrap4
django-fontawesome-5